update errorHandler, DELETE /api/organization route

+++ server/src/services/organizations.ts
@@ -1,9 +1,10 @@
 // /services/organizations.ts
 
 import Organization from "../models/organization";
+import User from "../models/user";
 import { IOrganization } from "../types/organization";
 import { IUser } from "../types/user";
-import { Document } from "mongoose";
+import { Document, Types } from "mongoose";
 
 export const getOrganization = async(id: string) => {
     const org = await Organization.findById(id);
@@ -17,13 +18,21 @@ export const addOrganization = async(data: IOrganization) => {
     return await Organization.create(data);
 };
 
-export const removeOrganization = async(id: string) => {
-    const organization = await Organization.findById(id);
+export const removeOrganization = async (id: string, user: IUser & Document) => {
+    if (user.organizationId.toString() !== id) return 'unauthorized';
 
+    const organization = await Organization.findById(id);
     if (!organization) return null;
 
-    return await organization.deleteOne();
+    //Deleting organization and their users parallely
+    await Promise.all([
+        organization.deleteOne(),
+        User.deleteMany({ organizationId: new Types.ObjectId(id) }),
+    ]);
+
+    return 'deleted';
 };
+
 
 export const updateOrganization = async (orgId: string, updates: IOrganization, user: (IUser & Document)) => {
     if (orgId !== user.organizationId.toString()) return 'unauthorized';
